Feature/138 fe profile invite for game modal (#150)
Implemented the invite for game modal in the profile view. All the
changes made in the interface are registered and you can see them being
consoled. Its not fully implemented because there's currently no
entrypoint.

**How to test?**

Open the invite for game modal and the console. Make changes to the game
setting and submit. confirm that the logs coincide with the changes.
experiment with different users.

---------

Co-authored-by: ahokcool <stein.alexander@posteo.de>
Co-authored-by: Alex <114107898+ahokcool@users.noreply.github.com>