translations done in the profile view (#163)
**What was implemented**

Translation on the dynamic inserted elements of the profile view.
Meaning that every text that is inserted using javascript is now
translated. If the text content comes directly from a HTML element the
translation is not done. This happens because me and @rajh-phuyal will
abstract those translations.

---
**How to test?**
open the profile view -> open web inspector -> go to Application ->
change the language in the `store` field

Co-authored-by: ahokcool <stein.alexander@posteo.de>