WC: Search (#253)
Added a search bar to the navigation bar with the ability to search for
users.
API will be documented shortly, for now its only functional.
Also its know that there is not "Not found" message and translations
related to search yet.

Test:
Its currently mounted on the nav bar and can search and go to the
profile of the user.