fixed #329 now the ranks should be send correctly (#346)
@Xico
so as the wiki says we  have the member rentries:
```json
"tournamentMember":
{
    "id": 13,
    "username": "rphuyal",
    "avatarUrl": "54c455d5-761b-46a2-80a2-7a557d9ec618.png",
    "state": "accepted",
    "playedGames": 0,
    "wonGames": 0,
    "winPoints": 0,
    "rank": 0
}
```

While the game is in setup phase the rank of ALL players will be 0.
Since the fe should not display a rank than, this shouldn't be a
problem.

Right BEFORE the state changes to ONGOING the backend:
1. send all tournamentMembers that haven't joined (with the `state`
`left`)
2. than it just randomly assign ranks to all remaining players
3. send all tournamentMembers with state `accepted`

Whenever a game finishes the backend will send all tournament members a
single messages again:

![image](https://github.com/user-attachments/assets/e753a6dd-84ab-4e60-ac85-b43bc60c61ef)