Play with AI (#420)
## AI Player Module

Create a game in home and start playing 🎉



https://github.com/user-attachments/assets/333ecd45-1624-4354-9d3d-a3dfae2a5b55




In this major module, the objective is to incorporate an AI player into
the game. Notably, the use of the **A* algorithm** is not permitted for
this task. Key features and goals include:

- Develop an AI opponent that provides a challenging and engaging
gameplay experience for users.

- The AI must replicate human behavior, meaning that in your AI
implementation, you must simulate keyboard input. The constraint here is
that the AI can only refresh its view of the game once per second,
requiring it to anticipate bounces and other actions.

> **ℹ️ The AI must utilize power-ups if you have chosen to implement the
Game customization options module.**

- Implement AI logic and decision-making processes that enable the AI
player to make intelligent and strategic moves.

- Explore alternative algorithms and techniques to create an effective
AI player without relying on A*.

- Ensure that the AI adapts to different gameplay scenarios and user
interactions.

> **⚠️ Attention: You will need to explain in detail how your AI
functions during your evaluation. Creating an AI that does nothing is
strictly prohibited; it must have the capability to win occasionally.**