[ONGOING] Enhancement/438 fe translation on router (#440)
### **The translation for everything static is now happening in the
router! :D**

**Important notes**
- The translations are done on all view and modals.
- only the static text is translated this way. This means that only text
that is on the **DOM after DOM insertion will be translated**. All other
text needs to be translated in the respective `configs.js` file, using
the `translate` function from `locale.js`.

**Major changes**

1. the `loadTranslations` action is now loading the `translations.js`
file from the modals directories. This files don't yet exists, they are
currently being done by @ahokcool.
2. @ahokcool is also creating `staticTranslations.json` files for each
view and modal. This files contains a json that maps the id of an
element that needs translation and the translation name for the
`translations.json` file.

**How does it works**
1. The translations are loaded from the `translations.json` files into
the local storage.
2. After DOM insertion the router loads the `staticTranslation.json`
file from the view that its routing.
3. it iterates through the map and translates the elements.
4. repeats the same process with the modals for the routing view