protection for endpoint: /user/username-exists/<str:search>/ (#491)
So my tester (`./deploy.sh test`) noticed, that we don't protect the
other request methods on this endpoint. We have to do this since we
can't use BaseAuthenticatedView here. So I added them manually. If u run
the tester before and after the fix u will see the difference for tests:
166 and 167 :v:

When reviewed u can merge this branch :v:

This is the current best case for the tester:
If you have time you can check the test cases in the
[table](https://docs.google.com/spreadsheets/d/11uhG7xBOxUAyQIAAKTghT1Bh5gUYxqQ2fPpRNr6fjds/edit?gid=0#gid=0)
to see why they fail. (Test Number = Line Number)


![image](https://github.com/user-attachments/assets/ec7f99d0-3a28-4c8f-b172-302f9d8339d1)