Fix/501 fe bug 404 cannot set properties of null (#507)
Me and @ahokcool had to do a bit of a refactoring to fix this issue.
There were a lot of errors that were causing the bug in the issue #501.
The main problem was that the `setTranslations mutationListener` was
routing aswell. This meant that the view was being routed twice on
login, to two different views.

---------

Co-authored-by: Alexander Stein <stein.alexander@posteo.de>